Using <zrtp-hash/> element check the integrity of the ZRTP key exchange.
<zrtp-hash/> element (as defined in XEP-0262) contains hash value of the
peer ZRTP Hello packet and the version of ZRTP protocol. With
setReceivedSignaledZRTPVersion, setReceivedSignaledZRTPHashValue
push the hash and version received from Jingle description down to
ZrtpControl and ZRTPTransformEngine.

Later, when the ZRTP Hello packet is received, the callback
ZRTPTransformEngine.sendInfo compares the hash value and version
of the ZRTP Hello packet with the hash value and version from Jingle
description.

fix race condition in AbstractSrtpControl.getTransformEngine